Deprecate publicapi "reflow"
reflow has two meanings in here. One is the publicapi (or an internal method)
telling us that we should rerender the mathquill to handle measurements. we don't
need that anymore.

The other is an event that bubbles up saying that mathquill _did_ reflow because it changed in some way.
We do need that.

This turns the .reflow() public API method into a no-op (it already is).
It gets rid of two places that we were doing extra work because we thought we might need to
trigger an extra reflow.

It also gets rid of a now-useless test.

There's possibly more cleanup that we can do as well around the 'reflow' event, but that's for later.